# specs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Stellar Scout API",
		"version": "1.6.1",
		"summary": "Agent-native data layer for the Stellar ecosystem — projects, builders, hackathons, SCF, audits, research.",
...
		"contact": {
			"name": "Stellar Light",
			"url": "https://stellarlight.xyz/scout",
			"email": "support@stellarlight.xyz"
		},
		"license": {
			"name": "MIT",
			"identifier": "MIT"
		}
	},
	"servers": [
		{
			"url": "https://stellarlight.xyz",
			"description": "Production"
		}
	],
...
	"paths": {
...
		"/api/skills": {
			"get": {
				"operationId": "listSkills",
				"tags": [
					"Skills"
				],
				"summary": "List AI skills for Stellar builders",
				"description": "Catalog of installable Stellar **AI skills/tools** — a unified, filterable list merging SDF's 7 official skills.stellar.org skills, Stellarlight + lumenloop + trusted third-party curated entries, and approved community submissions. Each entry carries an `install` command, `kind` (skill-md | mcp-server | sdk | cli | agent-kit | tool), `source`, repo/homepage/docs, and `meta.counts.bySource`; filter by `source` or `kind`. **Use when:** 'what Stellar AI skills / MCP servers / SDKs can I install', 'is there a skill for soroban / anchors / payments', or matching a builder's idea to the right installable tool. **Not for:** the full SKILL.md text or install details of ONE named skill → use get_skill; built/funded products in the ecosystem (not installable agent skills) → use search_projects; GitHub source repos → use search_repos; docs/standards/how-to knowledge → use search_research.",
				"parameters": [
					{
						"name": "source",
						"in": "query",
						"description": "Filter by source",
						"schema": {
							"type": "string",
							"enum": [
								"sdf",
								"stellarlight",
								"lumenloop",
								"external",
								"community"
							]
						}
					},
					{
						"name": "kind",
						"in": "query",
						"description": "Filter by skill kind",
						"schema": {
							"type": "string",
							"enum": [
								"skill-md",
								"mcp-server",
								"sdk",
								"cli",
								"agent-kit",
								"tool"
							]
						}
					}
				],
				"responses": {
					"200": {
						"description": "Skills catalog",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}